# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.21
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled and accept either
    a dashboard session cookie or a trusted internal `OMNIROUTE_MANAGEMENT_TOKEN` Bearer token.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/keys/{id}:
    get:
      tags: [API Keys]
      summary: Get API key
      description: |
        Requires management authentication when login protection is enabled. Trusted
        internal backends such as SLAI may use `Authorization: Bearer <OMNIROUTE_MANAGEMENT_TOKEN>`.
      security:
        - ManagementSessionAuth: []
        - ManagementBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: API key detail
        "401":
          $ref: "#/components/responses/ManagementAuthenticationRequired"
        "403":
          $ref: "#/components/responses/ManagementInvalidToken"
    patch:
      tags: [API Keys]
      summary: Update API key
      description: |
        Requires management authentication when login protection is enabled. Trusted
        internal backends such as SLAI may use `Authorization: Bearer <OMNIROUTE_MANAGEMENT_TOKEN>`.
      security:
        - ManagementSessionAuth: []
        - ManagementBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: API key updated
        "401":
          $ref: "#/components/responses/ManagementAuthenticationRequired"
        "403":
          $ref: "#/components/responses/ManagementInvalidToken"
    delete:
      tags: [API Keys]
      summary: Delete API key
      description: |
        Requires management authentication when login protection is enabled. Trusted
        internal backends such as SLAI may use `Authorization: Bearer <OMNIROUTE_MANAGEMENT_TOKEN>`.
      security:
        - ManagementSessionAuth: []
        - ManagementBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Key deleted
        "401":
          $ref: "#/components/responses/ManagementAuthenticationRequired"
        "403":
          $ref: "#/components/responses/ManagementInvalidToken"
...
  /api/settings:
    get:
      tags: [Settings]
      summary: Get application settings
      responses:
        "200":
          description: Current settings
    patch:
      tags: [Settings]
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated settings
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the OmniRoute dashboard
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie for protected management routes
    ManagementBearerAuth:
      type: http
      scheme: bearer
      description: Trusted server-to-server management token from `OMNIROUTE_MANAGEMENT_TOKEN`
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    ManagementAuthenticationRequired:
      description: Authentication required for management routes
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            error:
              message: Authentication required
              type: invalid_request
            requestId: 3f9f6f5a-509a-4b35-b0a7-2d2d99d73a01
    ManagementInvalidToken:
      description: Bearer management token is present but invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            error:
              message: Invalid management token
              type: invalid_request
            requestId: 1b6a6ff8-d60c-4900-8d0a-25f81749f0a3
...
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            details:
              description: Optional additional error details
        requestId:
          type: string
          format: uuid
...